Renaming SimsNode to AbstractSimsNode and SimsNode HeapedSimsNode.
