+++ cpp_low_index/cpp_src/abstractSimsNode.h
@@ -1,0 +1,61 @@
+#ifndef LOW_INDEX_ABSTRACT_SIMS_NODE_H
+#define LOW_INDEX_ABSTRACT_SIMS_NODE_H
+
+#include "coveringSubgraph.h"
+
+namespace low_index {
+
+class AbstractSimsNode : public CoveringSubgraph
+{
+public:
+
+    // Name new_relators_lift to make clear that relators_may_lift always
+    // has to be called with the same relators?
+    bool relators_lift(const std::vector<Relator> &relators) const;
+    bool relators_may_lift(const std::vector<Relator> &relators);
+    bool may_be_minimal() const;
+
+    unsigned int num_relators() const { return _num_relators; }
+
+protected:
+    AbstractSimsNode(RankType rank,
+                     DegreeType max_degree,
+                     unsigned int num_relators);
+
+    AbstractSimsNode(const AbstractSimsNode &other);
+
+    struct _MemoryLayout
+    {
+        _MemoryLayout(const AbstractSimsNode &node);
+
+        static constexpr size_t outgoing_offset = 0;
+        size_t incoming_offset;
+        size_t lift_indices_offset;
+        size_t lift_vertices_offset;
+        size_t size;
+    };
+
+    void _apply_memory_layout(const _MemoryLayout &layout,
+                              uint8_t * memory);
+    void _initialize_memory();
+    void _copy_memory(const AbstractSimsNode &other);
+
+private:
+    bool _relator_may_lift(
+        const Relator &relator,
+        size_t n,
+        DegreeType v);
+
+    bool _may_be_minimal(DegreeType basepoint) const;
+
+    const unsigned int _num_relators;
+
+public:
+    size_t _memory_size;
+    RelatorLengthType *_lift_indices;
+    DegreeType *_lift_vertices;
+};
+
+} // Namespace low_index
+
+#endif
