+++ cpp_low_index/cpp_src/simsTree.h
@@ -1,7 +1,7 @@
 #ifndef LOW_INDEX_SIMS_TREE_H
 #define LOW_INDEX_SIMS_TREE_H
 
-#include "heapedSimsNode.h"
+#include "simsNode.h"
 
 #include <atomic>
 
@@ -18,33 +18,33 @@ public:
         const std::vector<Relator> &short_relators,
         const std::vector<Relator> &long_relators);
 
-    std::vector<HeapedSimsNode> list(
+    std::vector<SimsNode> list(
         size_t bloom_size,
         unsigned int thread_num) const;
 
 private:
     SimsTree(
-        const HeapedSimsNode &root,
+        const SimsNode &root,
         const std::vector<Relator> &short_relators,
         const std::vector<Relator> &long_relators);
 
     void _recurse(
         const StackedSimsNode &n,
-        std::vector<HeapedSimsNode> *nodes) const;
-    std::vector<HeapedSimsNode> _bloom(size_t n) const;
+        std::vector<SimsNode> *nodes) const;
+    std::vector<SimsNode> _bloom(size_t n) const;
 
-    std::vector<HeapedSimsNode> _list_single_threaded() const;
-    std::vector<HeapedSimsNode> _list_multi_threaded(
+    std::vector<SimsNode> _list_single_threaded() const;
+    std::vector<SimsNode> _list_multi_threaded(
         size_t bloom_size, unsigned int thread_num) const;
 
     void _thread_worker(
-        const std::vector<HeapedSimsNode> &branches,
+        const std::vector<SimsNode> &branches,
         std::atomic_size_t * index,
-        std::vector<std::vector<HeapedSimsNode>> * nested_result) const;
+        std::vector<std::vector<SimsNode>> * nested_result) const;
     
     const std::vector<Relator> _short_relators;
     const std::vector<Relator> _long_relators;
-    const HeapedSimsNode _root;
+    const SimsNode _root;
 };
 
 } // Namespace low_index
