+++ cpp_low_index/cpp_src/wrapModule.cpp
@@ -3,8 +3,8 @@
 namespace low_index {
 
 void addCoveringSubgraph(pybind11::module_ &m);
+void addAbstractSimsNode(pybind11::module_ &m);
 void addSimsNode(pybind11::module_ &m);
-void addHeapedSimsNode(pybind11::module_ &m);
 void addSimsTree(pybind11::module_ &m);
 
 }
@@ -20,7 +20,7 @@ A C++ reimplementation of `low_index`.
 )pbdoc";
 
     addCoveringSubgraph(m);
+    addAbstractSimsNode(m);
     addSimsNode(m);
-    addHeapedSimsNode(m);
     addSimsTree(m);
 }
